+++ trading_system/patterns/__init__.py
@@ -1,5 +1,12 @@
 """Chart-pattern scan + multi-timeframe bias."""
 
+from trading_system.patterns.coverage import (
+    CONTINUATION_PATTERN_PREFIXES,
+    HIGH_PRIORITY_NAMES,
+    coverage_report,
+    htf_bb_entry_mode,
+    is_continuation_pattern,
+)
 from trading_system.patterns.detectors import (
     DetectedPattern,
     WEAK_PATTERNS,
@@ -12,16 +19,21 @@ from trading_system.patterns.mtf import combine_htf_votes, ltf_turn, macd_htf_bi
 from trading_system.patterns.pivots import find_pivots, is_near
 
 __all__ = [
+    "CONTINUATION_PATTERN_PREFIXES",
     "DetectedPattern",
+    "HIGH_PRIORITY_NAMES",
     "WEAK_PATTERNS",
     "best_reversal",
-    "measure_target",
-    "pattern_opposes_htf",
-    "scan_patterns",
     "combine_htf_votes",
+    "coverage_report",
+    "find_pivots",
+    "htf_bb_entry_mode",
+    "is_continuation_pattern",
+    "is_near",
     "ltf_turn",
     "macd_htf_bias",
+    "measure_target",
+    "pattern_opposes_htf",
     "resample_ohlcv",
-    "find_pivots",
-    "is_near",
+    "scan_patterns",
 ]